Fix participant endpoint

+++ src/main/java/pl/edu/agh/travelagencyapp/controller/ParticipantController.java
@@ -11,9 +11,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import pl.edu.agh.travelagencyapp.exception.InvalidParticipantException;
 import pl.edu.agh.travelagencyapp.exception.ResourceNotFoundException;
 import pl.edu.agh.travelagencyapp.model.Participant;
 import pl.edu.agh.travelagencyapp.model.Reservation;
+import pl.edu.agh.travelagencyapp.model.Trip;
+import pl.edu.agh.travelagencyapp.model.User;
 import pl.edu.agh.travelagencyapp.repository.ParticipantRepository;
 import pl.edu.agh.travelagencyapp.repository.ReservationRepository;
 
@@ -50,11 +53,38 @@ public class ParticipantController {
         return ResponseEntity.ok().body(new Participant(participant));
     }
 
+    @GetMapping("/participants/{id}/reservation")
+    public ResponseEntity<Reservation> getParticipantReservation(@PathVariable(value = "id") Long participantId)
+            throws ResourceNotFoundException {
+        Participant participant = participantRepository.findById(participantId)
+                .orElseThrow(() -> new ResourceNotFoundException("Participant with id " + participantId + " not found!"));
+        return ResponseEntity.ok().body(new Reservation(participant.getReservation()));
+    }
+
+    @GetMapping("/participants/{id}/user")
+    public ResponseEntity<User> getParticipantUser(@PathVariable(value = "id") Long participantId)
+            throws ResourceNotFoundException {
+        Participant participant = participantRepository.findById(participantId)
+                .orElseThrow(() -> new ResourceNotFoundException("Participant with id " + participantId + " not found!"));
+        return ResponseEntity.ok().body(new User(participant.getReservation().getUser()));
+    }
+
+    @GetMapping("/participants/{id}/trip")
+    public ResponseEntity<Trip> getParticipantTrip(@PathVariable(value = "id") Long participantId)
+            throws ResourceNotFoundException {
+        Participant participant = participantRepository.findById(participantId)
+                .orElseThrow(() -> new ResourceNotFoundException("Participant with id " + participantId + " not found!"));
+        return ResponseEntity.ok().body(new Trip(participant.getReservation().getTrip()));
+    }
+
     @PostMapping("/participants")
     public Participant createParticipant(@RequestBody Participant participant, @RequestParam(value = "reservationId") Long reservationId)
-            throws ResourceNotFoundException {
+            throws ResourceNotFoundException, InvalidParticipantException {
         Reservation reservation = reservationRepository.findById(reservationId)
                 .orElseThrow(() -> new ResourceNotFoundException("Reservation with id " + reservationId + " not found!"));
+
+        if(reservation.getParticipants().size() >= reservation.getNumberOfParticipants())
+            throw new InvalidParticipantException("Cannot add participant. Number of participants is as declared");
 
         participant.setReservation(reservation);
 
@@ -76,9 +106,12 @@ public class ParticipantController {
 
     @DeleteMapping("/participants/{id}")
     public Map<String, Boolean> deleteParticipant(@PathVariable(value = "id") Long participantId)
-            throws ResourceNotFoundException {
+            throws ResourceNotFoundException, InvalidParticipantException {
         Participant participant = participantRepository.findById(participantId)
                 .orElseThrow(() -> new ResourceNotFoundException("Participant with id " + participantId + " not found!"));
+
+        if(!participant.getReservation().getStatus().equals("N"))
+            throw new InvalidParticipantException("Cannot delete participant. Reservation status is not NEW");
 
         this.participantRepository.delete(participant);
 
@@ -88,6 +121,7 @@ public class ParticipantController {
         return response;
     }
 
+    //HIDDEN
     @DeleteMapping("/participants")
     public Map<String, Boolean> deleteAllParticipants() {
